feat: move to custom-functions
