+++ src/tests/fixtures/test_sputnik/test_sputnik_extension.did
@@ -1,1 +1,2 @@
-service : { hello : () -> () query; world : () -> (principal) }
+type InputArgs = record { value : principal };
+service : { world_world : (InputArgs) -> () query }
